@@ -1,11 +1,10 @@
 name: multi_navigator_bottom_bar
 description: Helps you to build multi-navigator bottom navigation bar more easily.
 version: 0.0.8
-author: showang <showang730@gmail.com>
 homepage: https://github.com/showang/flutter_multi_navigator_bottom_bar
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
