@@ -7,10 +7,10 @@ class BottomBarTab {
 
   /// Called when the tab is tapped and the navigator of the tab in on
   /// the first page.
-  final VoidCallback initialPageTappedCallback;
+  final VoidCallback? initialPageTappedCallback;
 
   /// Name of the initial page.
-  final String initialPageName;
+  final String? initialPageName;
 
   /// Builds the icon for the tab.
   ///
@@ -20,15 +20,13 @@ class BottomBarTab {
   /// Builds the active icon for the tab.
   ///
   /// You can return a widget like [Icon] here.
-  final WidgetBuilder tabActiveIconBuilder;
+  final WidgetBuilder? tabActiveIconBuilder;
 
   /// Builds the title for the tab.
-  ///
-  /// You can return a widget like [Text] here.
-  final WidgetBuilder tabTitleBuilder;
+  final String? tabTitle;
 
   /// Builds the routes for the tab.
-  final WidgetBuilder routePageBuilder;
+  final WidgetBuilder? routePageBuilder;
 
   /// The observers for the navigator.
   final List<NavigatorObserver> observers;
@@ -38,14 +36,14 @@ class BottomBarTab {
 
   /// Creates a new instance.
   BottomBarTab({
-    @required this.initialPageBuilder,
-    @required this.tabIconBuilder,
+    required this.initialPageBuilder,
+    required this.tabIconBuilder,
     this.initialPageTappedCallback,
     this.tabActiveIconBuilder,
     this.initialPageName,
-    this.tabTitleBuilder,
+    this.tabTitle,
     this.routePageBuilder,
-    this.observers,
-    GlobalKey<NavigatorState> navigatorKey,
+    this.observers = const [],
+    GlobalKey<NavigatorState>? navigatorKey,
   }) : this.navigatorKey = navigatorKey ?? GlobalKey<NavigatorState>();
 }
